requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.13
chardet==3.0.4
Django==2.1
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-redis==4.9.0
idna==2.7
kombu==4.2.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-memcached==1.59
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
SQLAlchemy-Utils==0.33.3
urllib3==1.23
vine==1.1.4
